+++ Realm/RLMApp.mm
@@ -288,6 +288,11 @@ NSError *RLMAppErrorToNSError(realm::app::AppError const& appError) {
 static NSMutableDictionary *s_apps = [NSMutableDictionary new];
 static std::mutex& s_appMutex = *new std::mutex();
 
++ (NSArray *)appIds {
+    std::lock_guard<std::mutex> lock(s_appMutex);
+    return s_apps.allKeys;
+}
+
 + (void)resetAppCache {
     std::lock_guard<std::mutex> lock(s_appMutex);
     [s_apps removeAllObjects];